requirements.txt
huggingface-hub
matplotlib
wandb
webdataset
accelerate>=0.20.3
einops>=0.6.1
importlib-metadata==6.7.0
iopath
numpy
omegaconf>=2.3.0
packaging>=23.2
Pillow==9.5.0
PyYAML==6.0.1
scipy
tabulate
tensorboardX
torch-fidelity
torchinfo>=1.8.0
tqdm
webdataset>=0.2.57
zipp
termcolor
clean-fid
tensorflow